+++ Editor/Core/Documentation/DocumentationHelpers.cs
@@ -27,13 +27,13 @@ namespace ThunderKit.Core.Documentation
 
             AssetDatabase.Refresh();
 
-            #if UNITY_6000_5_OR_NEWER
+            #if UNITY_6000_4_OR_NEWER
             Action<EntityId, string, string> action =
             #else
             Action<int, string, string> action =
             #endif
                 (
-                    #if UNITY_6000_5_OR_NEWER
+                    #if UNITY_6000_4_OR_NEWER
                     EntityId instanceId,
                     #else
                     int instanceId,