+++ Editor/Core/Windows/PipelineLogWindow.cs
@@ -71,13 +71,13 @@ namespace ThunderKit.Core.Windows
 
 
         [OnOpenAsset]
-        #if UNITY_6000_5_OR_NEWER
+        #if UNITY_6000_4_OR_NEWER
         public static bool OnOpen(EntityId entityId, int line)
         #else
         public static bool OnOpen(int instanceId, int line)
         #endif
         {
-            #if UNITY_6000_5_OR_NEWER
+            #if UNITY_6000_4_OR_NEWER
             var asset = EditorUtility.EntityIdToObject(entityId);
             #else
             var asset = EditorUtility.InstanceIDToObject(instanceId);